Delete tools/chocolateyBeforeModify.ps1
From mod TheCakeIsNaOH:
The before modify script should not have uninstall code. If you need to remove an old version(s) of the software before an upgrade, that logic/code should be in the install script. Given that the previous versions of the package did not have a before modify at all, I'd suggest removing it entirely, as it looks like the installer can handle installing on top of old software versions.